requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==6.0.2
psycopg2-binary==2.9.11
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
